requirements.txt
faiss-cpu==1.12.0
fastembed==0.7.3
feedparser==6.0.11
pandas==2.2.3
numpy==2.1.3
tqdm==4.66.4
fastapi==0.115.0
uvicorn==0.30.6
python-dateutil==2.9.0
requests==2.32.3
jinja2==3.1.4
xgboost==2.1.2
torch==2.1.0
scikit-learn==1.3.2
matplotlib==3.8.2
